+++ src/server/api/v1/__init__.py
@@ -1,0 +1,6 @@
+"""
+Copyright (c) 2024, 2025, Oracle and/or its affiliates.
+Licensed under the Universal Permissive License v1.0 as shown at http://oss.oracle.com/licenses/upl.
+"""
+
+from . import chat, databases, embed, models, oci, probes, prompts, testbed, settings, selectai
