+++ src/server/api/v1/models.py
@@ -17,7 +17,7 @@ logger = logging_config.logging.getLogger("endpoints.v1.models")
 auth = APIRouter()
 
 
-@auth.get("/v1/models/api", description="Get support model APIs", response_model=list)
+@auth.get("/api", description="Get support model APIs", response_model=list)
 async def models_list_api(
     model_type: Optional[schema.ModelTypeType] = Query(None),
 ) -> list[schema.Model]:
@@ -31,7 +31,7 @@ async def models_list_api(
         return list()
 
 
-@auth.get("/v1/models", description="Get all models", response_model=list[schema.Model])
+@auth.get("/", description="Get all models", response_model=list[schema.Model])
 async def models_list(
     model_type: Optional[schema.ModelTypeType] = Query(None),
 ) -> list[schema.Model]:
@@ -42,7 +42,7 @@ async def models_list(
     return models_ret
 
 
-@auth.get("/v1/models/{name:path}", description="Get a single model", response_model=schema.Model)
+@auth.get("/{name:path}", description="Get a single model", response_model=schema.Model)
 async def models_get(name: schema.ModelNameType) -> schema.Model:
     """List a specific model"""
     logger.debug("Received models_get - name: %s", name)
@@ -52,10 +52,10 @@ async def models_get(name: schema.ModelNameType) -> schema.Model:
     except ValueError as ex:
         raise HTTPException(status_code=404, detail=str(ex)) from ex
 
-    return models_ret[0]
+    return models_ret
 
 
-@auth.patch("/v1/models/{name:path}", description="Update a model", response_model=schema.Model)
+@auth.patch("/{name:path}", description="Update a model", response_model=schema.Model)
 async def models_update(name: schema.ModelNameType, payload: schema.Model) -> schema.Model:
     """Update a model"""
     logger.debug("Received models_update - name: %s; payload: %s", name, payload)
@@ -70,7 +70,7 @@ async def models_update(name: schema.ModelNameType, payload: schema.Model) -> sc
     return models.get_model(model_name=name)
 
 
-@auth.post("/v1/models", description="Create a model", response_model=schema.Model)
+@auth.post("/", description="Create a model", response_model=schema.Model)
 async def models_create(payload: schema.Model) -> schema.Model:
     """Update a model"""
     logger.debug("Received model_create - payload: %s", payload)
@@ -81,7 +81,7 @@ async def models_create(payload: schema.Model) -> schema.Model:
         raise HTTPException(status_code=409, detail=str(ex)) from ex
 
 
-@auth.delete("/v1/models/{name:path}", description="Delete a model")
+@auth.delete("/{name:path}", description="Delete a model")
 async def models_delete(name: schema.ModelNameType) -> JSONResponse:
     """Delete a model"""
     logger.debug("Received models_delete - name: %s", name)
