+++ src/server/api/v1/oci.py
@@ -2,12 +2,15 @@
 Copyright (c) 2024, 2025, Oracle and/or its affiliates.
 Licensed under the Universal Permissive License v1.0 as shown at http://oss.oracle.com/licenses/upl.
 """
-# spell-checker:ignore selectai
+# spell-checker:ignore selectai ocid
 
 from fastapi import APIRouter, HTTPException, Header
 from fastapi.responses import JSONResponse
 
-from server.api.core import oci, embed, models
+from server.api.core import models
+import server.api.core.oci as core_oci
+from server.api.util import embed
+import server.api.util.oci as util_oci
 
 import common.schema as schema
 import common.logging_config as logging_config
@@ -17,18 +20,18 @@ logger = logging_config.logging.getLogger("endpoints.v1.oci")
 auth = APIRouter()
 
 
-@auth.get("/v1/oci", description="View OCI Configuration", response_model=list[schema.OracleCloudSettings])
+@auth.get("/", description="View OCI Configuration", response_model=list[schema.OracleCloudSettings])
 async def oci_list() -> list[schema.OracleCloudSettings]:
     """List OCI Configuration"""
     logger.debug("Received oci_list")
     try:
-        return oci.get_oci()
+        return core_oci.get_oci()
     except ValueError as ex:
-        raise HTTPException(status_code=404, detail=str(ex)) from ex
+        raise HTTPException(status_code=404, detail=f"OCI: {str(ex)}.") from ex
 
 
 @auth.get(
-    "/v1/oci/{auth_profile}",
+    "/{auth_profile}",
     description="View OCI Profile Configuration",
     response_model=schema.OracleCloudSettings,
 )
@@ -36,13 +39,13 @@ async def oci_get(auth_profile: schema.OCIProfileType) -> schema.OracleCloudSett
     """List OCI Configuration"""
     logger.debug("Received oci_get - auth_profile: %s", auth_profile)
     try:
-        return oci.get_oci(auth_profile=auth_profile)
+        return core_oci.get_oci(auth_profile=auth_profile)
     except ValueError as ex:
-        raise HTTPException(status_code=404, detail=str(ex)) from ex
+        raise HTTPException(status_code=404, detail=f"OCI: {str(ex)}.") from ex
 
 
 @auth.get(
-    "/v1/oci/compartments/{auth_profile}",
+    "/compartments/{auth_profile}",
     description="Get OCI Compartments",
     response_model=dict,
 )
@@ -50,12 +53,12 @@ async def oci_list_compartments(auth_profile: schema.OCIProfileType) -> dict:
     """Return a list of compartments"""
     logger.debug("Received oci_list_compartments - auth_profile: %s", auth_profile)
     oci_config = await oci_get(auth_profile=auth_profile)
-    compartments = oci.get_compartments(oci_config)
+    compartments = util_oci.get_compartments(oci_config)
     return compartments
 
 
 @auth.get(
-    "/v1/oci/buckets/{compartment_ocid}/{auth_profile}",
+    "/buckets/{compartment_ocid}/{auth_profile}",
     description="Get OCI Object Storage buckets in Compartment OCID",
     response_model=list,
 )
@@ -64,12 +67,12 @@ async def oci_list_buckets(auth_profile: schema.OCIProfileType, compartment_ocid
     logger.debug("Received oci_list_buckets - auth_profile: %s; compartment_ocid: %s", auth_profile, compartment_ocid)
     compartment_obj = schema.OracleResource(ocid=compartment_ocid)
     oci_config = await oci_get(auth_profile=auth_profile)
-    buckets = oci.get_buckets(compartment_obj.ocid, oci_config)
+    buckets = util_oci.get_buckets(compartment_obj.ocid, oci_config)
     return buckets
 
 
 @auth.get(
-    "/v1/oci/objects/{bucket_name}/{auth_profile}",
+    "/objects/{bucket_name}/{auth_profile}",
     description="Get OCI Object Storage buckets objects",
     response_model=list,
 )
@@ -77,12 +80,12 @@ async def oci_list_bucket_objects(auth_profile: schema.OCIProfileType, bucket_na
     """Return a list of bucket objects; Validate OCID using Pydantic class"""
     logger.debug("Received oci_list_bucket_objects - auth_profile: %s; bucket_name: %s", auth_profile, bucket_name)
     oci_config = await oci_get(auth_profile=auth_profile)
-    objects = oci.get_bucket_objects(bucket_name, oci_config)
+    objects = util_oci.get_bucket_objects(bucket_name, oci_config)
     return objects
 
 
 @auth.patch(
-    "/v1/oci/{auth_profile}",
+    "/{auth_profile}",
     description="Update, Test, Set as Default OCI Configuration",
     response_model=schema.OracleCloudSettings,
 )
@@ -92,9 +95,9 @@ async def oci_profile_update(
     """Update OCI Configuration"""
     logger.debug("Received oci_update - auth_profile: %s; payload %s", auth_profile, payload)
     try:
-        namespace = oci.get_namespace(payload)
-    except oci.OciException as ex:
-        raise HTTPException(status_code=401, detail=str(ex)) from ex
+        namespace = util_oci.get_namespace(payload)
+    except util_oci.OciException as ex:
+        raise HTTPException(status_code=401, detail=f"OCI: {str(ex)}.") from ex
 
     oci_config = await oci_get(auth_profile=auth_profile)
     try:
@@ -128,7 +131,7 @@ async def oci_profile_update(
 
 
 @auth.post(
-    "/v1/oci/objects/download/{bucket_name}/{auth_profile}",
+    "/objects/download/{bucket_name}/{auth_profile}",
     description="Download files from Object Storage",
 )
 async def oci_download_objects(
@@ -148,7 +151,7 @@ async def oci_download_objects(
     # Files should be placed in the embedding folder
     temp_directory = embed.get_temp_directory(client, "embedding")
     for object_name in request:
-        oci.get_object(temp_directory, object_name, bucket_name, oci_config)
+        util_oci.get_object(temp_directory, object_name, bucket_name, oci_config)
 
     downloaded_files = [f.name for f in temp_directory.iterdir() if f.is_file()]
     return JSONResponse(status_code=200, content=downloaded_files)
