+++ tests/server/test_endpoints_databases.py
@@ -56,7 +56,9 @@ class TestEndpoints:
         assert default_db["password"] is None
         assert default_db["tcp_connect_timeout"] == 5
         assert default_db["user"] is None
-        assert default_db["vector_stores"] is None
+        assert default_db["vector_stores"] == []
+        assert default_db["selectai"] is False
+        assert default_db["selectai_profiles"] == []
         assert default_db["wallet_location"] is None
         assert default_db["wallet_password"] is None
