+++ tests/server/test_endpoints_oci.py
@@ -58,15 +58,15 @@ def _mock_init_client():
     mock_client.get_namespace.return_value.data = "test_namespace"
     mock_client.get_object.return_value.data.raw.stream.return_value = [b"fake-data"]
 
-    with patch("server.utils.oci.init_client", return_value=mock_client):
+    with patch("server.api.util.oci.init_client", return_value=mock_client):
         yield mock_client
 
 
 @pytest.fixture(name="mock_get_compartments")
 def _mock_get_compartments():
-    """Mock server_oci.get_compartments"""
+    """Mock get_compartments"""
     with patch(
-        "server.utils.oci.get_compartments",
+        "server.api.util.oci.get_compartments",
         return_value={
             "compartment1": "ocid1.compartment.oc1..aaaaaaaagq33tv7wzyrjar6m5jbplejbdwnbjqfqvmocvjzsamuaqnkkoubq",
             "compartment1 / test": "ocid1.compartment.oc1..aaaaaaaaut53mlkpxo6vpv7z5qlsmbcc3qpdjvjzylzldtb6g3jia",
@@ -80,7 +80,7 @@ def _mock_get_compartments():
 def _mock_get_buckets():
     """Mock server_oci.get_buckets"""
     with patch(
-        "server.utils.oci.get_buckets",
+        "server.api.util.oci.get_buckets",
         return_value=["bucket1", "bucket2", "bucket3"],
     ) as mock:
         yield mock
@@ -90,7 +90,7 @@ def _mock_get_buckets():
 def _mock_get_bucket_objects():
     """Mock server_oci.get_bucket_objects"""
     with patch(
-        "server.utils.oci.get_bucket_objects",
+        "server.api.util.oci.get_bucket_objects",
         return_value=["object1.pdf", "object2.md", "object3.txt"],
     ) as mock:
         yield mock
@@ -99,14 +99,14 @@ def _mock_get_bucket_objects():
 @pytest.fixture(name="mock_get_namespace")
 def _mock_get_namespace():
     """Mock server_oci.get_namespace"""
-    with patch("server.utils.oci.get_namespace", return_value="test_namespace") as mock:
+    with patch("server.api.util.oci.get_namespace", return_value="test_namespace") as mock:
         yield mock
 
 
 @pytest.fixture(name="mock_get_object")
 def _mock_get_object():
     """Mock get_object to return a fake file path"""
-    with patch("server.utils.oci.get_object") as mock:
+    with patch("server.api.util.oci.get_object") as mock:
 
         def side_effect(temp_directory, object_name):
             fake_file = temp_directory / object_name
@@ -141,7 +141,7 @@ class TestEndpoints:
         assert data["auth_profile"] == "DEFAULT"
         response = client.get("/v1/oci/TEST", headers=auth_headers["valid_auth"])
         assert response.status_code == 404
-        assert response.json() == {"detail": "OCI: Profile TEST not found."}
+        assert response.json() == {"detail": "OCI: profile 'TEST' not found."}
 
     def test_oci_list_compartments(self, client, auth_headers, mock_get_compartments):
         """List OCI Compartments"""
@@ -153,10 +153,10 @@ class TestEndpoints:
 
             # Test TEST profile
             mock_get.return_value.status_code = 404
-            mock_get.return_value.json.return_value = {"detail": "OCI: Profile TEST not found."}
+            mock_get.return_value.json.return_value = {"detail": "OCI: profile 'TEST' not found."}
             response = client.get("/v1/oci/compartments/TEST", headers=auth_headers["valid_auth"])
             assert response.status_code == 404
-            assert response.json() == {"detail": "OCI: Profile TEST not found."}
+            assert response.json() == {"detail": "OCI: profile 'TEST' not found."}
 
     def test_oci_list_buckets(self, client, auth_headers, mock_get_buckets):
         """List OCI Buckets"""
@@ -169,12 +169,12 @@ class TestEndpoints:
 
             # Test TEST profile
             mock_get.return_value.status_code = 404
-            mock_get.return_value.json.return_value = {"detail": "OCI: Profile TEST not found."}
+            mock_get.return_value.json.return_value = {"detail": "OCI: profile 'TEST' not found."}
             response = client.get(
                 "/v1/oci/buckets/ocid1.compartment.oc1..aaaaaaaa/TEST", headers=auth_headers["valid_auth"]
             )
             assert response.status_code == 404
-            assert response.json() == {"detail": "OCI: Profile TEST not found."}
+            assert response.json() == {"detail": "OCI: profile 'TEST' not found."}
 
     def test_oci_list_bucket_objects(self, client, auth_headers, mock_get_bucket_objects):
         """List OCI Bucket Objects"""
@@ -185,10 +185,10 @@ class TestEndpoints:
 
             # Test TEST profile
             mock_get.return_value.status_code = 404
-            mock_get.return_value.json.return_value = {"detail": "OCI: Profile TEST not found."}
+            mock_get.return_value.json.return_value = {"detail": "OCI: profile 'TEST' not found."}
             response = client.get("/v1/oci/objects/bucket1/TEST", headers=auth_headers["valid_auth"])
             assert response.status_code == 404
-            assert response.json() == {"detail": "OCI: Profile TEST not found."}
+            assert response.json() == {"detail": "OCI: profile 'TEST' not found."}
 
     test_cases = [
         pytest.param("DEFAULT", "", 422, id="empty_payload"),
