+++ tests/server/test_endpoints_settings.py
@@ -49,12 +49,12 @@ class TestEndpoints:
         settings = response.json()
 
         # Verify the response contains the expected structure
-        assert settings["client"] == "default"
-        assert "ll_model" in settings
-        assert "prompts" in settings
-        assert "vector_search" in settings
-        assert "selectai" in settings
-        assert "oci" in settings
+        assert settings["client_settings"]["client"] == "default"
+        assert "ll_model" in settings["client_settings"]
+        assert "prompts" in settings["client_settings"]
+        assert "vector_search" in settings["client_settings"]
+        assert "selectai" in settings["client_settings"]
+        assert "oci" in settings["client_settings"]
 
     def test_settings_get_nonexistent_client(self, client, auth_headers):
         """Test getting settings for a non-existent client"""
@@ -80,18 +80,18 @@ class TestEndpoints:
         response = client.get("/v1/settings", headers=auth_headers["valid_auth"], params={"client": new_client})
         assert response.status_code == 200
         new_client_settings = response.json()
-        assert new_client_settings["client"] == new_client
+        assert new_client_settings["client_settings"]["client"] == new_client
 
         # Remove the client key to compare the rest
-        del default_settings["client"]
-        del new_client_settings["client"]
+        del default_settings["client_settings"]["client"]
+        del new_client_settings["client_settings"]["client"]
         assert default_settings == new_client_settings
 
     def test_settings_create_existing_client(self, client, auth_headers) -> None:
         """Test creating settings for an existing client"""
         response = client.post("/v1/settings", headers=auth_headers["valid_auth"], params={"client": "default"})
         assert response.status_code == 409
-        assert response.json() == {"detail": "Client: default already exists."}
+        assert response.json() == {"detail": "Settings: client default already exists."}
 
     def test_settings_update(self, client, auth_headers):
         """Test updating settings for a client"""
@@ -149,7 +149,7 @@ class TestEndpoints:
         response = client.patch(
             "/v1/settings",
             headers=auth_headers["valid_auth"],
-            json=default_settings,
+            json=default_settings["client_settings"],
             params={"client": "server"},
         )
         assert response.status_code == 200
@@ -157,8 +157,8 @@ class TestEndpoints:
         new_server_settings = response.json()
         assert old_server_settings != new_server_settings
 
-        del new_server_settings["client"]
-        del default_settings["client"]
+        del new_server_settings["client_settings"]["client"]
+        del default_settings["client_settings"]["client"]
         assert new_server_settings == default_settings
 
     def test_settings_update_nonexistent_client(self, client, auth_headers):
@@ -172,7 +172,7 @@ class TestEndpoints:
             params={"client": "nonexistent_client"},
         )
         assert response.status_code == 404
-        assert response.json() == {"detail": "Client: nonexistent_client not found."}
+        assert response.json() == {"detail": "Settings: client nonexistent_client not found."}
 
     @pytest.mark.parametrize("app_server", ["/tmp/settings.json"], indirect=True)
     def test_user_supplied_settings(self, app_server):
